refactor(diffs): centralize lazy renderer runtime
