+++ extensions/diffs/src/browser.runtime.ts
